requirements.txt
aiohttp==3.7.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.4.0
cachetools==4.2.2
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.121.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==2.10
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
metaapi-cloud-copyfactory-sdk==6.1.0
metaapi-cloud-metastats-sdk==3.2.1
metaapi-cloud-risk-management-sdk==1.2.1
metaapi-cloud-sdk==20.9.0
multidict==6.0.2
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
prettytable==3.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==3.14.2
python-socketio==4.6.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
rfc3986==1.5.0
schedule==1.2.1
six==1.16.0
sniffio==1.2.0
starlette==0.50.0
tornado==6.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==4.2
urllib3==1.25.11
uvicorn==0.38.0
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.1.2
yarl==1.7.2
